Move login/signup links into user account modal
